[ch73704] dedup test utils (#69)
