+++ test-utils/include/test-utils/client.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <launchdarkly/api.h>
+
+struct LDClient *makeTestClient();
+
+struct LDClient *makeOfflineClient();
