+++ test-utils/include/test-utils/store.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <launchdarkly/api.h>
+
+#include <store.h>
+
+void runSharedStoreTests(struct LDStore *(*const prepareEmptyStore)());
