+++ test-utils/src/flags.c
@@ -1,4 +1,4 @@
-#include <launchdarkly/api.h>
+#include "test-utils/flags.h"
 
 #include "assertion.h"
