+++ tests/test-event-processor.c
@@ -8,21 +8,8 @@
 #include "store.h"
 #include "client.h"
 
-#include "util-flags.h"
-
-static struct LDClient *
-makeOfflineClient()
-{
-    struct LDConfig *config;
-    struct LDClient *client;
-
-    LD_ASSERT(config = LDConfigNew("api_key"));
-    LDConfigSetOffline(config, true);
-
-    LD_ASSERT(client = LDClientInit(config, 0));
-
-    return client;
-}
+#include "test-utils/flags.h"
+#include "test-utils/client.h"
 
 static void
 testConstructAndFree()
