+++ tests/test-store-memory.c
@@ -3,7 +3,7 @@
 #include "assertion.h"
 #include "utility.h"
 
-#include "util-store.h"
+#include "test-utils/store.h"
 
 static struct LDStore *
 prepareEmptyStore()
